+++ src/ScottPlot4/ScottPlot.Tests/PlottableRenderTests/Annotation.cs
@@ -15,7 +15,7 @@ namespace ScottPlotTests.PlottableRenderTests
             var plt = new ScottPlot.Plot(400, 300);
 
             // start with default settings
-            var pa = new ScottPlot.Plottable.Annotation() { Label = "Hello", X = 10, Y = 10 };
+            var pa = new ScottPlot.Plottable.Annotation() { Label = "Hello" };
             pa.Font.Size = 36;
             plt.Add(pa);
             var bmp1 = TestTools.GetLowQualityBitmap(plt);
@@ -41,7 +41,7 @@ namespace ScottPlotTests.PlottableRenderTests
             var plt = new ScottPlot.Plot(400, 300);
 
             // start with default settings
-            var pa = new ScottPlot.Plottable.Annotation() { Label = "Hello", X = 10, Y = 10 };
+            var pa = new ScottPlot.Plottable.Annotation() { Label = "Hello" };
             plt.Add(pa);
             var bmp1 = TestTools.GetLowQualityBitmap(plt);
 
@@ -66,7 +66,7 @@ namespace ScottPlotTests.PlottableRenderTests
             var plt = new ScottPlot.Plot(400, 300);
 
             // start with default settings
-            var pa = new ScottPlot.Plottable.Annotation() { Label = "Hello", X = 10, Y = 10 };
+            var pa = new ScottPlot.Plottable.Annotation() { Label = "Hello" };
             plt.Add(pa);
             var bmp1 = TestTools.GetLowQualityBitmap(plt);
 
@@ -91,7 +91,7 @@ namespace ScottPlotTests.PlottableRenderTests
             var plt = new ScottPlot.Plot(400, 300);
 
             // start with default settings
-            var pa = new ScottPlot.Plottable.Annotation() { Label = "Hello", X = 10, Y = 10, BorderColor = System.Drawing.Color.Gray };
+            var pa = new ScottPlot.Plottable.Annotation() { Label = "Hello", BorderColor = System.Drawing.Color.Gray };
             plt.Add(pa);
             var bmp1 = TestTools.GetLowQualityBitmap(plt);
 
@@ -116,7 +116,7 @@ namespace ScottPlotTests.PlottableRenderTests
             var plt = new ScottPlot.Plot(400, 300);
 
             // start with default settings
-            var pa = new ScottPlot.Plottable.Annotation() { Label = "Hello", X = 10, Y = 10 };
+            var pa = new ScottPlot.Plottable.Annotation() { Label = "Hello" };
             plt.Add(pa);
             var bmp1 = TestTools.GetLowQualityBitmap(plt);